+++ src/net/offbeatpioneer/intellij/plugins/grav/module/builder/GravProjectGeneratorUtil.java
@@ -11,10 +11,13 @@ import com.intellij.openapi.roots.ModuleRootManager;
 import com.intellij.openapi.ui.MessageType;
 import com.intellij.openapi.ui.popup.Balloon;
 import com.intellij.openapi.ui.popup.JBPopupFactory;
+import com.intellij.openapi.vfs.VfsUtil;
 import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.openapi.vfs.VirtualFileEvent;
 import com.intellij.openapi.wm.StatusBar;
 import com.intellij.openapi.wm.WindowManager;
 import com.intellij.ui.awt.RelativePoint;
+import net.offbeatpioneer.intellij.plugins.grav.helper.FileCreateUtil;
 import net.offbeatpioneer.intellij.plugins.grav.module.tasks.InstallDevtoolsPlugin;
 import net.offbeatpioneer.intellij.plugins.grav.module.wizard.CopyFileVisitor;
 import net.offbeatpioneer.intellij.plugins.grav.project.settings.GravProjectSettings;
@@ -22,8 +25,11 @@ import org.jetbrains.annotations.NotNull;
 
 import java.io.File;
 import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 
 import static com.intellij.ide.projectView.actions.MarkRootActionBase.findContentEntry;
 
@@ -49,6 +55,10 @@ public class GravProjectGeneratorUtil {
                         model.commit();
                         test.refresh(false, true);
                         src1.refresh(false, true);
+                    } else {
+                        findContentEntry(model, src).addSourceFolder(src, false);
+                        model.commit();
+                        src.refresh(false, true);
                     }
 
                 } catch (IOException e) {
@@ -57,27 +67,28 @@ public class GravProjectGeneratorUtil {
             }
         });
 
-        File f = new File(module.getModuleFilePath());
-        String tmp = null;
-        try {
-            tmp = f.getCanonicalFile().getParent();
-            if (new File(tmp).isDirectory() && !new File(tmp).getName().equals(module.getName())) {
-                tmp = new File(tmp).getParent();
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        VirtualFile tmp = FileCreateUtil.getParentDirectory(module.getModuleFile(), module.getName());
         if (tmp == null) return false;
-        Path targetPath;
-        if (!settings.withSrcDirectory) {
-            targetPath = new File(tmp).toPath();
-        } else {
-            targetPath = new File(new File(tmp) + File.separator + "src").toPath();
+
+        URI targetPath;
+        try {
+            if (!settings.withSrcDirectory) {
+                targetPath = new URI("file:///" + tmp.getPath());
+            } else {
+                VirtualFile tmp1 = tmp.findFileByRelativePath("src");
+                if (tmp1 != null)
+                    targetPath = new URI("file:///" + tmp1.getPath());
+                else
+                    return false;
+            }
+        } catch (URISyntaxException e) {
+            e.printStackTrace();
+            return false;
         }
 
-        Task.Backgroundable installDevtools = new InstallDevtoolsPlugin(project, "Installing Devtools Plugin", new File(targetPath.toUri()));
+        Task.Backgroundable installDevtools = new InstallDevtoolsPlugin(project, "Installing Devtools Plugin", new File(targetPath));
 
-        Path finalTargetPath = targetPath;
+        URI finalTargetPath = targetPath;
         Task.Backgroundable t = new Task.Backgroundable(project, "Copying Grav SDK to Module Folder") {
 
             @Override
@@ -93,7 +104,7 @@ public class GravProjectGeneratorUtil {
 
             public void run(@NotNull ProgressIndicator progressIndicator) {
                 try {
-                    Files.walkFileTree(new File(settings.gravInstallationPath).toPath(), new CopyFileVisitor(finalTargetPath));
+                    Files.walkFileTree(new File(settings.gravInstallationPath).toPath(), new CopyFileVisitor(new File(finalTargetPath).toPath()));
                 } catch (IOException e) {
                     e.printStackTrace();
                 }